cinema：新增自动判断请求 ip 是否来自境外，如果是则自动添加进数据库中并在本次请求以及下次请求拦截该请求。添加了是否在服务器重启时重置视频缩略图数据以及是否开启视频变更监控

+++ src/com/jeffrey/listener/ContextListener.java
@@ -10,58 +10,85 @@ import javax.servlet.ServletContext;
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.ArrayList;
+import java.util.Properties;
 
 public class ContextListener implements ServletContextListener {
 
     private static final ThumbnailService THUMBNAIL_SERVICE = new ThumbnailServiceImpl();
     private static final ArrayList<File> LISTS = new ArrayList<>();
+    private static boolean initData;
+    private static boolean monitorFolder;
+
+    static {
+        InputStream is = null;
+        try {
+            is = ContextListener.class.getClassLoader().getResourceAsStream("conf/root.properties");
+            Properties prop = new Properties();
+            prop.load(is);
+            initData = Boolean.parseBoolean(prop.getProperty("initData"));
+            monitorFolder = Boolean.parseBoolean(prop.getProperty("monitorFolder"));
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        if (is != null) {
+            try {
+                is.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+    }
 
 
     @Override
     public void contextInitialized(ServletContextEvent sce) {
+
+        ServletContext servletContext = sce.getServletContext();
+        // 监听目录
+        File target = new File(servletContext.getRealPath("/static/video"));
+        // 缩略图存放目录
+        String thumbnailPath = servletContext.getRealPath("/static/thumbnail/");
+
         new Thread(() -> {
-            // 对表数据进行充重置
-            THUMBNAIL_SERVICE.truncateData();
 
-            ServletContext servletContext = sce.getServletContext();
+            if (initData) {
 
-            // 监听目录
-            File target = new File(servletContext.getRealPath("/static/video"));
+                // 对表数据进行充重置
+                THUMBNAIL_SERVICE.truncateData();
 
-            // 查找监听目录内的所有 mp4 视频
-            findVideo(target);
+                // 查找监听目录内的所有 mp4 视频
+                findVideo(target);
 
-            String thumbnailPath = servletContext.getRealPath("/static/thumbnail/");
 
-            File[] files = new File(thumbnailPath).listFiles();
-            if (files != null) {
-                for (File item : files) {
-                    if (item.delete()) {
-                        System.out.println("移除文件：" + item);
+                File[] files = new File(thumbnailPath).listFiles();
+                if (files != null) {
+                    for (File item : files) {
+                        if (item.delete()) {
+                            System.out.println("移除文件：" + item);
+                        }
                     }
                 }
+
+                ThumbnailDataManagement.createThumbnail(LISTS, thumbnailPath);
             }
 
-            ThumbnailDataManagement.createThumbnail(LISTS, thumbnailPath);
+            if (monitorFolder) {
+                FileAlterationObserver observer = new FileAlterationObserver(target);
 
-            FileAlterationObserver observer = new FileAlterationObserver(target);
+                observer.addListener(new MonitorVideoModify(thumbnailPath, servletContext.getRealPath("/")));
 
-            observer.addListener(new MonitorVideoModify(thumbnailPath, servletContext.getRealPath("/")));
+                FileAlterationMonitor monitor = new FileAlterationMonitor(5000, observer);
 
-            FileAlterationMonitor monitor = new FileAlterationMonitor(5000, observer);
-
-            try {
-                monitor.start();
-            } catch (Exception e) {
-                e.printStackTrace();
+                try {
+                    monitor.start();
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
             }
         }).start();
-    }
-
-    @Override
-    public void contextDestroyed(ServletContextEvent sce) {
-
     }
 
     private static void findVideo(File file) {
@@ -78,4 +105,6 @@ public class ContextListener implements ServletContextListener {
             }
         }
     }
+    @Override
+    public void contextDestroyed(ServletContextEvent sce) {}
 }
